Package shuffling for settings and cycles
