+++ controller/capabilities_test.go
@@ -3,6 +3,7 @@ package controller
 import (
 	"bytes"
 	"encoding/json"
+	"github.com/reef-pi/reef-pi/controller/settings"
 	"github.com/reef-pi/reef-pi/controller/utils"
 	"net/http"
 	"testing"
@@ -19,7 +20,7 @@ func Test_Capabilities(t *testing.T) {
 	if err := tr.Do("GET", "/api/capabilities", new(bytes.Buffer), nil); err != nil {
 		t.Error("Failed to fetch capabilities ror:", err)
 	}
-	var resp Capabilities
+	var resp settings.Capabilities
 	if err := json.NewDecoder(buf).Decode(&resp); err != nil {
 		t.Error(err)
 	}
