+++ controller/health_test.go
@@ -1,6 +1,7 @@
 package controller
 
 import (
+	"github.com/reef-pi/reef-pi/controller/settings"
 	"github.com/reef-pi/reef-pi/controller/utils"
 	"testing"
 )
@@ -11,7 +12,7 @@ func TestHealthChecker(t *testing.T) {
 		t.Error(err)
 	}
 	telemetry := utils.TestTelemetry()
-	h := NewHealthChecker(1, HealthCheckNotify{}, telemetry, store)
+	h := NewHealthChecker(1, settings.HealthCheckNotify{}, telemetry, store)
 	h.Notify.Enable = true
 	h.check()
 }
